requirements.txt
flask>=2.0.1
python-dotenv>=0.19.0
supabase>=0.7.0
python-jose>=3.3.0
flask-cors>=3.0.10
